requirements.txt
# =============================================================================
# Credit Risk Prediction - Production Requirements
# =============================================================================

# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0

# Deep Learning (Optional)
torch>=2.0.0

# Model Explainability
shap>=0.43.0
lime>=0.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Data Quality & Profiling
missingno>=0.5.2
great-expectations>=0.17.0
pandera>=0.17.0

# MLOps & Experiment Tracking
mlflow>=2.8.0
dagshub>=0.3.0

# AWS SDK
boto3>=1.28.0
botocore>=1.31.0
awswrangler>=3.4.0
sagemaker>=2.190.0

# API & Web Framework
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.4.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Database & Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redshift-connector>=2.0.0
pyarrow>=13.0.0
s3fs>=2023.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
moto>=4.2.0
hypothesis>=6.87.0

# Code Quality
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0
pre-commit>=3.4.0

# Logging & Monitoring
structlog>=23.1.0
python-json-logger>=2.0.0
prometheus-client>=0.17.0

# Utilities
tqdm>=4.66.0
joblib>=1.3.0
click>=8.1.0
rich>=13.5.0

# Jupyter & Notebooks
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.1.0
nbformat>=5.9.0
papermill>=2.4.0
